rename args file to constants
