+++ extension/src/cli/executor.test.ts
@@ -2,7 +2,7 @@ import { join } from 'path'
 import { EventEmitter } from 'vscode'
 import { Disposable, Disposer } from '@hediet/std/disposable'
 import { CliResult, CliStarted } from '.'
-import { Flag, GcPreserveFlag } from './args'
+import { Flag, GcPreserveFlag } from './constants'
 import { CliExecutor } from './executor'
 import { createProcess } from '../processExecution'
 import { getMockedProcess } from '../test/util/jest'
