+++ extension/src/cli/options.ts
@@ -1,5 +1,5 @@
 import { dirname } from 'path'
-import { Args } from './args'
+import { Args } from './constants'
 import { getProcessEnv } from '../env'
 import { joinEnvPath } from '../util/env'
