+++ extension/src/commands/internal.ts
@@ -2,7 +2,7 @@ import { commands } from 'vscode'
 import { Disposable } from '@hediet/std/disposable'
 import { RegisteredCliCommands, RegisteredCommands } from './external'
 import { ICli } from '../cli'
-import { Args } from '../cli/args'
+import { Args } from '../cli/constants'
 import { autoRegisteredCommands as CliExecutorCommands } from '../cli/executor'
 import { autoRegisteredCommands as CliReaderCommands } from '../cli/reader'
 import { autoRegisteredCommands as CliRunnerCommands } from '../cli/runner'
